backend/requirements.txt
# Core Flask Dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6

# WSGI Server
gunicorn==21.2.0
eventlet==0.35.2

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Machine Learning
torch>=2.6.0
numpy==1.26.2
scipy==1.11.4

# Computer Vision
opencv-python-headless==4.8.1.78
mediapipe==0.10.32

# Validation & Serialization
marshmallow==3.20.1
pydantic==2.5.2

# Security
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk[flask]==1.38.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
factory-boy==3.3.0
httpx==0.25.2
locust==2.20.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
bandit==1.7.6
isort==5.13.2

# Type stubs
types-redis==4.6.0.11
types-requests==2.31.0.10

# Utilities
python-dotenv==1.0.0
click==8.1.7
tenacity==8.2.3
